[PUBDEV-5332] Make downloadLogs method public
